Rename other non-hardcoded setup-mc or SetupMC to onboarding.
