+++ tests/e2e/utils/pages/onboarding/step-1-set-up-accounts.js
