+++ tests/e2e/utils/pages/onboarding/step-2-product-listings.js
