+++ tests/e2e/utils/pages/onboarding/step-3-complete-campaign.js
